Moved main into a seperate directory
Enable 2 'main' functions, for both a server, and CLI.
