+++ exec/cli/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/PossibleLlama/worklog/cli"
+)
+
+func main() {
+	cli.Execute()
+}
