--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/PossibleLlama/worklog/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
